+++ gdscript/src/main/kotlin/gdscript/index/stub/GdClassVarDeclStub.kt
@@ -5,5 +5,6 @@ import gdscript.psi.GdClassVarDeclTl
 import gdscript.psi.types.GdDocumented
 
 interface GdClassVarDeclStub : StubElement<GdClassVarDeclTl>, GdDocumented {
-    fun name(): String;
+    fun name(): String
+    fun isStatic(): Boolean
 }