+++ gdscript/src/main/kotlin/gdscript/psi/GdPsiUtils.kt
@@ -54,6 +54,7 @@ object GdPsiUtils {
     @JvmStatic fun getPresentation(element: GdClassVarDeclTl): ItemPresentation = GdPresentationUtil.presentation(element)
     @JvmStatic fun getReturnType(element: GdClassVarDeclTl): String = PsiGdClassVarUtil.getReturnType(element)
     @JvmStatic fun isAnnotated(element: GdClassVarDeclTl, annotator: String): Boolean = PsiGdClassVarUtil.isAnnotated(element, annotator)
+    @JvmStatic fun isStatic(element: GdClassVarDeclTl): Boolean = PsiGdClassVarUtil.isStatic(element)
 
     /** Local variable */
     @JvmStatic fun getName(element: GdVarDeclSt): String = GdVarDeclStUtil.getName(element)